+++ .idea/workspace.xml
@@ -2,7 +2,27 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="43dc116f-520a-48a6-aa65-947acd2709fb" name="Default" comment="">
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/Basic/autograd_tutorial.ipynb" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/Basic/autograd_tutorial.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/Basic/main_PT.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/Basic/main_PT_autograd.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/Basic/main_PT_numpy_bridge.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/Basic/tensor_tutorial.ipynb" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/Basic/tensor_tutorial.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/Install.txt" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/NN/MNISTConvNet.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/NN/autograd_tutorial.ipynb" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/NN/cifar10_tutorial.ipynb" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/NN/cifar10_tutorial.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/NN/main_CNN_NN_PT.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/NN/neural_networks_tutorial.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/NN/tensor_tutorial.ipynb" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/RNN/RNN.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/(준비중)PyTorch/RNN/nn_tutorial.py" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/TF_version_upgrade_for_1.0_util/ReadMeFor_tf_upgrade.txt" afterPath="" />
+      <change type="DELETED" beforePath="$PROJECT_DIR$/TF_version_upgrade_for_1.0_util/tf_upgrade.py" afterPath="" />
       <change type="MODIFICATION" beforePath="$PROJECT_DIR$/.idea/workspace.xml" afterPath="$PROJECT_DIR$/.idea/workspace.xml" />
+      <change type="MODIFICATION" beforePath="$PROJECT_DIR$/TensorFlow/iPythonNotebook/Convolutioanl Neural Network/main_CNN_TF.ipynb" afterPath="$PROJECT_DIR$/TensorFlow/iPythonNotebook/Convolutioanl Neural Network/main_CNN_TF.ipynb" />
     </list>
     <ignored path="DeepLearningTutorials.iws" />
     <ignored path=".idea/workspace.xml" />
@@ -59,10 +79,10 @@
     </option>
   </component>
   <component name="ProjectFrameBounds">
-    <option name="x" value="1328" />
-    <option name="y" value="215" />
-    <option name="width" value="1280" />
-    <option name="height" value="731" />
+    <option name="x" value="1272" />
+    <option name="y" value="-8" />
+    <option name="width" value="1936" />
+    <option name="height" value="1056" />
   </component>
   <component name="ProjectView">
     <navigator currentView="ProjectPane" proportions="" version="1">
@@ -79,8 +99,6 @@
       <foldersAlwaysOnTop value="true" />
     </navigator>
     <panes>
-      <pane id="Scope" />
-      <pane id="Scratches" />
       <pane id="ProjectPane">
         <subPane>
           <PATH>
@@ -113,6 +131,8 @@
           </PATH>
         </subPane>
       </pane>
+      <pane id="Scope" />
+      <pane id="Scratches" />
     </panes>
   </component>
   <component name="PropertiesComponent">
@@ -382,16 +402,16 @@
     <servers />
   </component>
   <component name="ToolWindowManager">
-    <frame x="1328" y="215" width="1280" height="731" extended-state="0" />
+    <frame x="1272" y="-8" width="1936" height="1056" extended-state="6" />
     <editor active="false" />
     <layout>
-      <window_info id="Project" active="false" anchor="left" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="true" show_stripe_button="true" weight="0.22463769" sideWeight="0.5" order="0" side_tool="false" content_ui="combo" />
+      <window_info id="Project" active="true" anchor="left" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="true" show_stripe_button="true" weight="0.14699684" sideWeight="0.5" order="0" side_tool="false" content_ui="combo" />
       <window_info id="TODO" active="false" anchor="bottom" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="false" show_stripe_button="true" weight="0.33" sideWeight="0.5" order="6" side_tool="false" content_ui="tabs" />
       <window_info id="Event Log" active="false" anchor="bottom" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="false" show_stripe_button="true" weight="0.33" sideWeight="0.5" order="7" side_tool="true" content_ui="tabs" />
       <window_info id="Version Control" active="false" anchor="bottom" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="false" show_stripe_button="true" weight="0.33" sideWeight="0.5" order="7" side_tool="false" content_ui="tabs" />
       <window_info id="Python Console" active="false" anchor="bottom" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="false" show_stripe_button="true" weight="0.32998326" sideWeight="0.5" order="7" side_tool="false" content_ui="tabs" />
       <window_info id="Structure" active="false" anchor="left" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="false" show_stripe_button="true" weight="0.25" sideWeight="0.5" order="1" side_tool="false" content_ui="tabs" />
-      <window_info id="Terminal" active="true" anchor="bottom" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="true" show_stripe_button="true" weight="0.32663316" sideWeight="0.5" order="7" side_tool="false" content_ui="tabs" />
+      <window_info id="Terminal" active="false" anchor="bottom" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="true" show_stripe_button="true" weight="0.21149674" sideWeight="0.5" order="7" side_tool="false" content_ui="tabs" />
       <window_info id="Favorites" active="false" anchor="left" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="false" show_stripe_button="true" weight="0.33" sideWeight="0.5" order="2" side_tool="true" content_ui="tabs" />
       <window_info id="Cvs" active="false" anchor="bottom" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="false" show_stripe_button="true" weight="0.25" sideWeight="0.5" order="4" side_tool="false" content_ui="tabs" />
       <window_info id="Message" active="false" anchor="bottom" auto_hide="false" internal_type="DOCKED" type="DOCKED" visible="false" show_stripe_button="true" weight="0.33" sideWeight="0.5" order="0" side_tool="false" content_ui="tabs" />
