--- TF_version_upgrade_for_1.0_util/ReadMeFor_tf_upgrade.txt
@@ -1,4 +1,0 @@
-ReadMe for upgrade TF Version (1.0)
-
-
-python tf_upgrade.py --infile InFileNameHere.py --outfile OutFileNameHere.py
